brings in the user so we can link to their stats
