+++ apps/web/app/(base)/2024/dataFor2024.ts
@@ -744,6 +744,38 @@ const dataFor2024 = {
     { venue: "Hollywood Theater", count: 9 },
     { venue: "PICA", count: 9 },
   ],
+  emojis: [
+    "🤙",
+    "🤸",
+    "🚣🏽",
+    "🌀",
+    "🪰",
+    "🪐",
+    "🍿",
+    "👻",
+    "🍪",
+    "👹",
+    "✨",
+    "🥜",
+    "🤌",
+    "🧦",
+    "🐢",
+    "🥑",
+    "🐯",
+    "🧜",
+    "🐞",
+    "💐",
+    "🦋",
+    "🌋",
+    "🌚",
+    "🎈",
+    "🧩",
+    "💖",
+    "😎",
+    "🙏",
+    "⏳",
+    "🔜",
+  ],
 };
 
 export default dataFor2024;
